+++ packages/web-components/lib/components/document/src/document-capture-instructions/index.js
@@ -1,3 +1,4 @@
 import DocumentInstruction from './DocumentCaptureInstructions';
+import './DocumentCaptureInstructions.tsx';
 
 export default DocumentInstruction;